update: CRUD function for client
created client service and client controller class, edited client class to use Lombok annotations for getter and setter method, simplified client mapper interface using mapstruct

+++ src/main/java/com/peercolab/homeservicebarbers/repository/BarberRepository.java
@@ -3,5 +3,5 @@ package com.peercolab.homeservicebarbers.repository;
 import com.peercolab.homeservicebarbers.entity.Barber;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BarberRepository extends JpaRepository<Barber, Integer> {
+public interface BarberRepository extends JpaRepository<Barber, Long> {
 }
